rag_system/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
langchain
langchain-community
langchain-openai
faiss-cpu==1.8.0.post1
rank_bm25==0.2.2
pypdf==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
